refactor app init service to use app data endpoint
