+++ src/store/index.ts
@@ -1,9 +1,9 @@
-export * from './auth-state.interface';
-export * from './auth.state.actions';
-export * from './auth.state';
-export * from './feature-config.state.actions';
-export * from './feature-config.state';
-export * from './group.state.actions';
-export * from './group.state';
-export * from './user.state.actions';
-export * from './user.state';
+export * from "./auth-state.interface";
+export * from "./auth.state.actions";
+export * from "./auth.state";
+export * from "./feature-config.state.actions";
+export * from "./feature-config.state";
+export * from "./group.state.actions";
+export * from "./group.state";
+export * from "./user.state.actions";
+export * from "./user.state";
